+++ src/main/scala/smptk/registration/Regularizer.scala
@@ -4,13 +4,13 @@ import scala.language.higherKinds
 import TransformationSpace.ParameterVector
 import breeze.linalg.DenseVector
 
-trait Regularizer extends (ParameterVector => Float) { 
+trait Regularizer extends (ParameterVector => Double) { 
   
-  def takeDerivative(p : ParameterVector) : DenseVector[Float]
+  def takeDerivative(p : ParameterVector) : DenseVector[Double]
 }
 
 object RKHSNormRegularizer extends Regularizer {
-	def apply(alpha : ParameterVector) = alpha.norm(2).toFloat
+	def apply(alpha : ParameterVector) = alpha.norm(2)
   
   def takeDerivative(alpha : ParameterVector) = alpha
 }
