+++ src/test/scala/smptk/registration/RegistrationTest.scala
@@ -19,30 +19,28 @@ import breeze.plot._
 class ImageTest extends FunSpec with ShouldMatchers {
   describe("A 2D rigid landmark based registration") {
     it("can retrieve correct parameters") {
-      val points = IndexedSeq(CoordVector2D(0f, 0f), CoordVector2D(1f, 4f), CoordVector2D(2f, 0f))
+      val points = IndexedSeq(CoordVector2D(0., 0.), CoordVector2D(1., 4.), CoordVector2D(2., 0.))
 
-      val c = CoordVector2D(1f, 4/3f) 
-      for (angle <- (1 until 16).map(i => (math.Pi).toFloat / i)) { 
-	      val rotationParams = DenseVector( -angle )
-	      val transParams = DenseVector[Float](1f, 1.5f)
-	      val productParams = DenseVector.vertcat(transParams, rotationParams)
-	
-	
-	      val productSpace = RigidTransformationSpace2D(c)
-	
-	      val transformedPoints = points.map((pt: CoordVector2D[Float]) => productSpace(productParams)(pt))
-	
-	      val (optiTransfrom, p) = LandmarkRegistration.rigid2DLandmarkRegistration(points.zip(transformedPoints), c)
-	       
-	      val alignedPoints = points.map((pt: CoordVector2D[Float]) => optiTransfrom(pt))
-	      
-	      
-	      (transformedPoints(0)(0) should be(alignedPoints(0)(0) plusOrMinus 0.0001f))
-	      (transformedPoints(0)(1) should be(alignedPoints(0)(1) plusOrMinus 0.0001f))
-	      (transformedPoints(1)(0) should be(alignedPoints(1)(0) plusOrMinus 0.0001f))
-	      (transformedPoints(1)(1) should be(alignedPoints(1)(1) plusOrMinus 0.0001f))
-	      (transformedPoints(2)(0) should be(alignedPoints(2)(0) plusOrMinus 0.0001f))
-	      (transformedPoints(2)(1) should be(alignedPoints(2)(1) plusOrMinus 0.0001f))
+      val c = CoordVector2D(1., 4 / 3.)
+      for (angle <- (1 until 16).map(i => math.Pi / i)) {
+        val rotationParams = DenseVector(-angle)
+        val transParams = DenseVector[Double](1f, 1.5f)
+        val productParams = DenseVector.vertcat(transParams, rotationParams)
+
+        val productSpace = RigidTransformationSpace2D(c)
+
+        val transformedPoints = points.map((pt: CoordVector2D[Double]) => productSpace(productParams)(pt))
+
+        val (optiTransfrom, p) = LandmarkRegistration.rigid2DLandmarkRegistration(points.zip(transformedPoints), c)
+
+        val alignedPoints = points.map((pt: CoordVector2D[Double]) => optiTransfrom(pt))
+
+        (transformedPoints(0)(0) should be(alignedPoints(0)(0) plusOrMinus 0.0001))
+        (transformedPoints(0)(1) should be(alignedPoints(0)(1) plusOrMinus 0.0001))
+        (transformedPoints(1)(0) should be(alignedPoints(1)(0) plusOrMinus 0.0001))
+        (transformedPoints(1)(1) should be(alignedPoints(1)(1) plusOrMinus 0.0001))
+        (transformedPoints(2)(0) should be(alignedPoints(2)(0) plusOrMinus 0.0001))
+        (transformedPoints(2)(1) should be(alignedPoints(2)(1) plusOrMinus 0.0001))
       }
     }
   }
